refector(card.js) : cards.js에 있던 내용을 분리
